# docs/swagger/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Go E-Commerce REST API",
    "description": "API untuk aplikasi E-Commerce sederhana dengan caching dan event-driven architecture",
    "version": "1.0.0"
  },
  "servers": [
    { "description": "local", "url": "http://localhost:3000/api/v1" },
    { "description": "docker container", "url": "http://localhost:3001/api/v1" }
  ],
  "paths": {
...
    "/products": {
      "get": {
        "security": [{ "productauth": [] }],
        "tags": ["product api"],
        "summary": "get all products with pagination",
        "description": "Get paginated list of products. Data is cached in Redis for 10 minutes.",
        "operationId": "getAllProducts",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Page number (starts from 1)",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "example": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "description": "Number of items per page",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 10,
              "example": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved products",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 200
                    },
                    "status": {
                      "type": "string",
                      "example": "OK"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/product"
                      }
                    }
                  }
                },
                "example": {
                  "code": 200,
                  "status": "OK",
                  "data": [
                    {
                      "id": 1,
                      "name": "Laptop",
                      "price": 15000000,
                      "description": "High performance laptop",
                      "category_id": 1,
                      "img": "https://example.com/laptop.jpg"
                    },
                    {
                      "id": 2,
                      "name": "Mouse",
                      "price": 150000,
                      "description": "Wireless mouse",
                      "category_id": 2,
                      "img": "https://example.com/mouse.jpg"
                    }
                  ]
                }
              }
            },
            "headers": {
              "X-Cache-Status": {
                "description": "Indicates whether data was served from cache",
                "schema": {
                  "type": "string",
                  "enum": ["HIT", "MISS"]
                }
              }
            }
          },
          "400": {
            "description": "Invalid pagination parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/errorResponse"
                },
                "example": {
                  "code": 400,
                  "status": "BAD REQUEST",
                  "data": {
                    "error": "Invalid page parameter"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Missing or invalid token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/errorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/errorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "security": [{ "productauth": [] }],
        "tags": ["product api"],
        "description": "create new product",
        "operationId": "createproduct",
        "summary": "create new product",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "price", "category_id"],
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "Laptop Gaming"
                  },
                  "price": {
                    "type": "integer",
                    "minimum": 0,
                    "example": 15000000
                  },
                  "description": {
                    "type": "string",
                    "example": "High performance gaming laptop"
                  },
                  "category_id": {
                    "type": "integer",
                    "example": 1
                  },
                  "img": {
                    "type": "string",
                    "format": "uri",
                    "example": "https://example.com/laptop.jpg"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Product created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": { "type": "number", "example": 201 },
                    "status": { "type": "string", "example": "CREATED" },
                    "data": { "$ref": "#/components/schemas/product" }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/errorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/errorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/errorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/products/{id}": {
      "get": {
        "security": [{ "productauth": [] }],
        "tags": ["product api"],
        "summary": "get product by id",
        "description": "Get single product by ID. Data is cached in Redis for 10 minutes.",
        "operationId": "getProductById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Product ID",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "example": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Product found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": { "type": "number", "example": 200 },
                    "status": { "type": "string", "example": "OK" },
                    "data": { "$ref": "#/components/schemas/product" }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Product not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/errorResponse"
                },
                "example": {
                  "code": 404,
                  "status": "NOT FOUND",
                  "data": {
                    "error": "Product not found"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/errorResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "security": [{ "productauth": [] }],
        "tags": ["product api"],
        "summary": "update product by id",
        "description": "Update product and invalidate cache",
        "operationId": "updateProductById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "example": 1
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string" },
                  "price": { "type": "integer", "minimum": 0 },
                  "description": { "type": "string" },
                  "category_id": { "type": "integer" },
                  "img": { "type": "string", "format": "uri" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Product updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": { "type": "number", "example": 200 },
                    "status": { "type": "string", "example": "OK" },
                    "data": { "$ref": "#/components/schemas/product" }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Product not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/errorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "security": [{ "productauth": [] }],
        "tags": ["product api"],
        "summary": "delete product by id",
        "description": "Delete product and invalidate cache",
        "operationId": "deleteProductById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "example": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Product deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": { "type": "number", "example": 200 },
                    "status": { "type": "string", "example": "OK" },
                    "data": {
                      "type": "object",
                      "properties": {
                        "message": {
                          "type": "string",
                          "example": "Product deleted successfully"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Product not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/errorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "userauth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter JWT token (without 'Bearer' prefix)"
      },
      "productauth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter JWT token for product operations"
      },
      "transactionauth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter JWT token for transaction operations"
      }
    },
    "schemas": {
...
      "product": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "example": 1 },
          "name": { "type": "string", "example": "Laptop Gaming" },
          "price": { "type": "integer", "example": 15000000 },
          "description": { "type": "string", "example": "High performance gaming laptop" },
          "category_id": { "type": "integer", "example": 1 },
          "img": { "type": "string", "format": "uri", "example": "https://example.com/laptop.jpg" }
        }
      },
...
      "errorResponse": {
        "type": "object",
        "properties": {
          "code": { "type": "integer", "example": 400 },
          "status": { "type": "string", "example": "BAD REQUEST" },
          "data": {
            "type": "object",
            "properties": {
              "error": { "type": "string", "example": "Validation error" }
            }
          }
        }
      }
    }
  }
}